pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "purl2meta"
version = "0.1.0"
description = "Resolve Package URLs to registry metadata API URLs"
requires-python = ">=3.8"
license = "MIT"
dependencies = ["packageurl-python"]

[project.optional-dependencies]
fetch = ["httpx"]
test = ["pytest", "httpx"]

[tool.hatch.build.targets.wheel]
packages = ["src/purl2meta"]
